requirements.txt
numpy==1.17.3
opencv-contrib-python==4.1.2.30
python-dotenv==0.10.3
pathlib2==2.3.4
pytest==5.1.3
pytest-cov==2.7.1
pylint==2.3.1
python-json-logger==0.1.11
joblib==0.14.1
requests==2.23.0
tqdm==4.43.0